mkudffs: Fix creation of unalloc table
Creation of unalloc table was broken and the table tracked ICB with the
table as empty space which was then readily rewritten by the fileset
ICB resulting in corrupted filesystem.

Fix the problem by recording ICB with unalloc table as used space and
also add a check reporting error when too large filesystem with unalloc
table should be created instead of creating corrupted fs.

Signed-off-by: Jan Kara <jack@suse.com>