mkudffs: Check FLAG_EFE rather than udf_rev
I was getting a segmentation fault/core dump with this command line:

mkudffs --u8 --udfrev=0x0102 --blocksize=2048 --lvid="HDi-250 DVD" --strategy=4 \
	--media-type=dvdrw --space=unallocbitmap --ad=inicb --noefe /dev/sr0

The segfault was caused by a subscript of -1 while lookup for a descriptor
but the root cause was using extended file entries on a UDF-1.02 disc
where they are not supported. The check for version == 1.50 should have
been <= 1.50, but checking for FLAG_EFE is better since it also deals
with the --noefe option.

Signed-off-by: Steve Kenton <skenton@ou.edu>