mkudffs: Add new option --bootarea=preserve|erase to preserve or erase UDF boot area
In most cases UDF boot area (first 32kB of disk) is not used and therefore
it is better to erase it. By default when mkudffs formatting hard disk will
erase UDF boot area which cleanup headers of previous filesystems. So hard
disk would not be false positive detected as other filesystem which was
there before formatting to UDF.