mkudffs: work around chkdsk inability to analyze 4K media
See this thread for background:  https://lkml.org/lkml/2019/7/9/596 (ISSUE 2)

On media with 4096-byte blocks, Windows chkdsk looks for VRS components
at 2048-byte intervals instead of the 4096-byte interval required by
the standard, and treats the media as RAW if components are not found.

When formatting 4K media, insert dummy BOOT2 descriptors midway between
the BEA and NSR and between the NSR and TEA. This is a UDF nonconformity,
but one that appears to make chkdsk happy without introducing
incompatibilities.