mkudffs: adjust partition size to work around chkdsk bug
Windows chkdsk has a bug in its Unallocated Space bitmap analysis.
If the last block of a UDF partition falls in the middle of a bitmap byte,
chkdsk reports spurious errors if the bits in that byte that _don't_
correspond to UDF partition blocks are zero.

To improve interoperability, when formatting includes an
unallocated space bitmap, make sure the created UDF partition
is a multiple of 8 blocks.