cdrwtool: Revive USE_IMMED
Some drives get stuck when a FORMAT UNIT without Immed bit command lasts
about 500 seconds. Other drives behave as if the Immed bit was set even if
it is not set.
The current code for waiting after starting an asynchronous SCSI command
is only active if USE_IMMED is 1. It wrongly expects wait_cmd() to fill
its submitted sense data buffer and it determines the end of the
asynchronous command from the unreliable progress messages of the drive.

So replace functions print_completion_info() and sig_progress() by a new
function wait_for_unit_ready() and enable USE_IMMED.

Signed-off-by: Thomas Schmitt <scdbackup@gmx.net>