udftune: Add initial implementation
This patch adds a new udftools command/executable, `udftune`.

It implements two useful operations: "--mark-ro" and "--mark-rw", for
setting FSD and LVID Domain flags in a compatible UDF filesystem to
indicate either read-only or read/write characteristics.

The code borrows *heavily* from `udflabel`, which contains nearly all of
the required logic to modify a UDF target in this manner by itself.

Since `udflabel` is to be used for interacting with UDF label
information only, `udftune` was created in its current form.